% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{eventstudy}
\alias{eventstudy}
\title{Event Study Models and their implementation}
\description{
Wrapper for Difference in Differences with multiple periods and variation in
 treatment timing
}
