% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unc_att_gt_panel_het.R
\name{unc_att_gt_panel_het}
\alias{unc_att_gt_panel_het}
\title{Difference of ATT(g,t)'s across two subpopulations: Panel data and no covariates}
\usage{
unc_att_gt_panel_het(
  y_name,
  t_name,
  id_name,
  first_treat_name,
  het_name,
  weights_name = NULL,
  cluster_vars = NULL,
  comparison_group = c("never", "not_yet", "not_yet_all"),
  data,
  alp = 0.05,
  bstrap = TRUE,
  nboot = 999,
  print_details = TRUE,
  parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{y_name}{The name of the outcome variable}

\item{t_name}{The name of the column containing the time periods}

\item{id_name}{The individual (cross-sectional unit) id name}

\item{first_treat_name}{The name of the variable in \code{data} that
contains the first period when a particular observation is treated.
This should be a positive number for all observations in treated groups.
It defines which "group" a unit belongs to.  It should be 0 for units
in the untreated group.}

\item{het_name}{The name of the column containing the binary categories for heterogeneity}

\item{weights_name}{The name of the column containing the sampling weights.
If not set, all observations have same weight.}

\item{cluster_vars}{A vector of variables to cluster on.  At most, there
can be two variables (otherwise will throw an error) and one of these
must be the same as idname which allows for clustering at the individual
level.}

\item{comparison_group}{Which units to use the comparison group.
The default is \code{group="never"} which sets the comparison group
to be the group of units that never participate in the
treatment.  This group does not change across groups or
time periods. The other option is to set
\code{group="not_yet"}.  In this case, the comparison group
is set to the group of units that have not yet participated
in the treatment in that time period.  This includes all
never treated units, but it includes additional units that
eventually participate in the treatment, but have not
participated yet.}

\item{data}{The name of the data.frame that contains the data}

\item{alp}{the significance level, default is 0.05}

\item{bstrap}{Boolean for whether or not to compute standard errors using
the multiplier boostrap.  If standard errors are clustered, then one
must set \code{bstrap=TRUE}. Default is \code{TRUE} (in addition, cband
is also by default \code{TRUE} indicating that uniform confidence bands
will be returned.  If bstrap is \code{FALSE}, then analytical
standard errors are reported.}

\item{nboot}{The number of boostrap iterations to use.  The default is 999,
and this is only applicable if \code{bstrap=TRUE}.}

\item{print_details}{Boolean for showing detailed results or not}

\item{parallel}{Boolean for whether or not to use parallel processing
(not implemented yet)}

\item{cores}{The number of cores to use for parallel processing
(not implemented yet)}
}
\value{
an \code{MP} object containing all the results for group-time average
 treatment effects
}
\description{
\code{unc_att_gt_panel_het} computes the difference of average treatment effects across two subpopulations
in DID setups where there are more than two periods of data and
allowing for treatment to occur at different points in time. Here, we allow for different trends
between the two supopulations. See Marcus ans Sant'Anna (2020) for a detailed description.
}
\references{
Callaway, Brantly and Sant'Anna, Pedro H. C.. "Difference-in-Differences with Multiple Time Periods and an Application on the Minimum Wage and Employment." Working Paper <https://ssrn.com/abstract=3148250> (2019).

Marcus, Michelle and Sant'Anna, Pedro H. C.. "The Role of Parallel Trends in Event Study Settings: An Application to Environmental Economics." Working Paper <https://to_be_added.com> (2020).
}
